Resolve react/no-multi-comp Lint Issues
There were multiple components defined in the PortfolioLink component
file, this commit moves each component into its own definition file and
adds some snapshot tests for them.
