+++ src/component/portfoliolink/PortfolioLinkProps.ts
@@ -1,0 +1,7 @@
+export type PortfolioLinkProps = {
+  description: string;
+  url: string;
+  urlId: string;
+  issuerDescription?: string;
+  issuerUrl?: string;
+};
