+++ src/component/portfoliolink/__tests__/PortfolioLinkWithoutIssuer-test.js
@@ -1,0 +1,16 @@
+import PortfolioLinkWithIssuer from "../PortfolioLinkWithoutIssuer";
+import renderer from "react-test-renderer";
+
+it("renders to match snapshot", () => {
+  const tree = renderer
+    .create(
+      <PortfolioLinkWithIssuer
+        description="description Field"
+        url="https://example.com"
+        urlId="linkId Field"
+      />
+    )
+    .toJSON();
+
+  expect(tree).toMatchSnapshot();
+});
